Rename finder to locator
Current locator is just a proxy
