+++ lib/watir-webdriver/locators.rb
@@ -1,20 +1,19 @@
-require 'watir-webdriver/locators/element/finder'
 require 'watir-webdriver/locators/element/locator'
 require 'watir-webdriver/locators/element/selector_builder'
 require 'watir-webdriver/locators/element/validator'
 
-require 'watir-webdriver/locators/button/finder'
+require 'watir-webdriver/locators/button/locator'
 require 'watir-webdriver/locators/button/selector_builder'
 require 'watir-webdriver/locators/button/validator'
 
-require 'watir-webdriver/locators/cell/finder'
+require 'watir-webdriver/locators/cell/locator'
 require 'watir-webdriver/locators/cell/selector_builder'
 
-require 'watir-webdriver/locators/row/finder'
+require 'watir-webdriver/locators/row/locator'
 require 'watir-webdriver/locators/row/selector_builder'
 
-require 'watir-webdriver/locators/text_area/finder'
+require 'watir-webdriver/locators/text_area/locator'
 require 'watir-webdriver/locators/text_area/selector_builder'
 
-require 'watir-webdriver/locators/text_field/finder'
+require 'watir-webdriver/locators/text_field/locator'
 require 'watir-webdriver/locators/text_field/selector_builder'
