+++ lib/watir-webdriver/locators/cell/locator.rb
@@ -1,14 +1,14 @@
 module Watir
-  class Row
-    class Finder < Element::Finder
-      def find_all
+  class Cell
+    class Locator < Element::Locator
+      def locate_all
         find_all_by_multiple
       end
 
       private
 
       def by_id
-        nil # avoid this
+        nil
       end
     end
   end
